#5 - fixing SonarQube issues
